各ウィジェットの切り出し
