+++ flutter/test_project/lib/calendar.dart
@@ -1,3 +1,5 @@
+import 'package:flutter/material.dart';
+
 List makeDateList() {
   const dayNumList = [31, 28, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31];
   final monthList = List.generate(12, (index) => index + 1);
@@ -43,5 +45,72 @@ String weekdayFunction(int weekday) {
       return 'Sun';
     default:
       return 'error';
+  }
+}
+
+class Calendar extends StatelessWidget {
+  const Calendar({
+    super.key,
+    required this.dateList,
+  });
+
+  final List dateList;
+
+  @override
+  Widget build(BuildContext context) {
+    return Expanded(
+        child: ListView.builder(
+      itemCount: dateList.length,
+      itemBuilder: (context, index) {
+        final element = dateList[index];
+        final weekday = weekdayFunction(index);
+        var textColor = Colors.black;
+        if (weekday == 'Sat') {
+          textColor = Colors.blue;
+        }
+        if (weekday == 'Sun') {
+          textColor = Colors.red;
+        }
+        if (index == 0) {
+          return ListTile(
+            leading: Container(
+              width: 60,
+              height: 60,
+              decoration: const BoxDecoration(
+                  shape: BoxShape.circle, color: Colors.green),
+              child: const Center(
+                child: Text(
+                  'Today',
+                  style: TextStyle(
+                      fontWeight: FontWeight.w600,
+                      fontSize: 16,
+                      color: Colors.white),
+                ),
+              ),
+            ),
+            onTap: () => debugPrint('selected ${element.month}/${element.day}'),
+            title: Text(
+              '${weekday} ${element.month}/${element.day}',
+              style: TextStyle(color: textColor),
+            ),
+            subtitle: const Text('schedule:'),
+          );
+        } else {
+          return ListTile(
+            leading: const Icon(
+              Icons.arrow_forward,
+              size: 60,
+              color: Colors.blue,
+            ),
+            title: Text(
+              '${weekday} ${element.month}/${element.day}',
+              style: TextStyle(color: textColor),
+            ),
+            subtitle: const Text('schedule:'),
+            onTap: () => debugPrint('selected ${element.month}/${element.day}'),
+          );
+        }
+      },
+    ));
   }
 }
