+++ flutter/test_project/lib/main.dart
@@ -97,130 +97,3 @@ class MyHomePage extends StatelessWidget {
     );
   }
 }
-
-class Calendar extends StatelessWidget {
-  const Calendar({
-    super.key,
-    required this.dateList,
-  });
-
-  final List dateList;
-
-  @override
-  Widget build(BuildContext context) {
-    return Expanded(
-        child: ListView.builder(
-      itemCount: dateList.length,
-      itemBuilder: (context, index) {
-        final element = dateList[index];
-        final weekday = weekdayFunction(index);
-        var textColor = Colors.black;
-        if (weekday == 'Sat') {
-          textColor = Colors.blue;
-        }
-        if (weekday == 'Sun') {
-          textColor = Colors.red;
-        }
-        if (index == 0) {
-          return ListTile(
-            leading: Container(
-              width: 60,
-              height: 60,
-              decoration: const BoxDecoration(
-                  shape: BoxShape.circle, color: Colors.green),
-              child: const Center(
-                child: Text(
-                  'Today',
-                  style: TextStyle(
-                      fontWeight: FontWeight.w600,
-                      fontSize: 16,
-                      color: Colors.white),
-                ),
-              ),
-            ),
-            onTap: () => debugPrint('selected ${element.month}/${element.day}'),
-            title: Text(
-              '${weekday} ${element.month}/${element.day}',
-              style: TextStyle(color: textColor),
-            ),
-            subtitle: const Text('schedule:'),
-          );
-        } else {
-          return ListTile(
-            leading: const Icon(
-              Icons.arrow_forward,
-              size: 60,
-              color: Colors.blue,
-            ),
-            title: Text(
-              '${weekday} ${element.month}/${element.day}',
-              style: TextStyle(color: textColor),
-            ),
-            subtitle: const Text('schedule:'),
-            onTap: () => debugPrint('selected ${element.month}/${element.day}'),
-          );
-        }
-      },
-    ));
-  }
-}
-
-class StopWatch extends StatelessWidget {
-  const StopWatch({
-    super.key,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Row(
-      mainAxisAlignment: MainAxisAlignment.spaceAround,
-      children: [
-        Container(
-          width: 160,
-          child: ElevatedButton(
-            style: ElevatedButton.styleFrom(
-                backgroundColor: Colors.red,
-                shape: RoundedRectangleBorder(
-                    borderRadius: BorderRadius.circular(10))),
-            onPressed: () {
-              if (!start) {
-                setStart = true;
-                stopWatch(10);
-              }
-            },
-            child: const Row(
-              mainAxisAlignment: MainAxisAlignment.center,
-              children: [
-                Icon(
-                  Icons.start,
-                  color: Colors.white,
-                ),
-                Text('START', style: TextStyle(color: Colors.white))
-              ],
-            ),
-          ),
-        ),
-        Container(
-          width: 160,
-          child: ElevatedButton(
-            style: ElevatedButton.styleFrom(
-                backgroundColor: Colors.blue,
-                shape: RoundedRectangleBorder(
-                    borderRadius: BorderRadius.circular(10))),
-            onPressed: () => {setStart = false},
-            child: const Row(
-              mainAxisAlignment: MainAxisAlignment.center,
-              children: [
-                Icon(
-                  Icons.timelapse,
-                  color: Colors.white,
-                ),
-                Text('STOP', style: TextStyle(color: Colors.white))
-              ],
-            ),
-          ),
-        ),
-      ],
-    );
-  }
-}
