+++ flutter/test_project/lib/stop_watch.dart
@@ -21,3 +21,63 @@ Future<Duration> stopWatch(int maxMin) async {
   }
   return time;
 }
+
+class StopWatch extends StatelessWidget {
+  const StopWatch({
+    super.key,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Row(
+      mainAxisAlignment: MainAxisAlignment.spaceAround,
+      children: [
+        Container(
+          width: 160,
+          child: ElevatedButton(
+            style: ElevatedButton.styleFrom(
+                backgroundColor: Colors.red,
+                shape: RoundedRectangleBorder(
+                    borderRadius: BorderRadius.circular(10))),
+            onPressed: () {
+              if (!start) {
+                setStart = true;
+                stopWatch(10);
+              }
+            },
+            child: const Row(
+              mainAxisAlignment: MainAxisAlignment.center,
+              children: [
+                Icon(
+                  Icons.start,
+                  color: Colors.white,
+                ),
+                Text('START', style: TextStyle(color: Colors.white))
+              ],
+            ),
+          ),
+        ),
+        Container(
+          width: 160,
+          child: ElevatedButton(
+            style: ElevatedButton.styleFrom(
+                backgroundColor: Colors.blue,
+                shape: RoundedRectangleBorder(
+                    borderRadius: BorderRadius.circular(10))),
+            onPressed: () => {setStart = false},
+            child: const Row(
+              mainAxisAlignment: MainAxisAlignment.center,
+              children: [
+                Icon(
+                  Icons.timelapse,
+                  color: Colors.white,
+                ),
+                Text('STOP', style: TextStyle(color: Colors.white))
+              ],
+            ),
+          ),
+        ),
+      ],
+    );
+  }
+}
